build dockerfile taotao-cloud-gateway taotao-cloud-monitor taotao-cloud-file  taotao-cloud-job  taotao-cloud-auth taotao-cloud-sys  taotao-cloud-xxljob

+++ taotao-cloud-microservice/taotao-cloud-business/taotao-cloud-payment/taotao-cloud-payment-biz/src/main/java/com/taotao/cloud/payment/biz/jeepay/service/impl/MchAppService.java
@@ -1,40 +1,26 @@
-/*
- * Copyright (c) 2020-2030, Shuigedeng (981376577@qq.com & https://blog.taotaocloud.top/).
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package com.taotao.cloud.payment.biz.jeepay.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
-import com.taotao.cloud.payment.biz.jeepay.core.constants.ApiCodeEnum;
-import com.taotao.cloud.payment.biz.jeepay.core.constants.CS;
-import com.taotao.cloud.payment.biz.jeepay.core.entity.MchApp;
-import com.taotao.cloud.payment.biz.jeepay.core.entity.MchPayPassage;
-import com.taotao.cloud.payment.biz.jeepay.core.entity.PayInterfaceConfig;
-import com.taotao.cloud.payment.biz.jeepay.core.entity.PayOrder;
-import com.taotao.cloud.payment.biz.jeepay.core.exception.BizException;
-import com.taotao.cloud.payment.biz.jeepay.core.utils.StringKit;
-import com.taotao.cloud.payment.biz.jeepay.service.mapper.MchAppMapper;
+import com.jeequan.jeepay.core.constants.ApiCodeEnum;
+import com.jeequan.jeepay.core.constants.CS;
+import com.jeequan.jeepay.core.entity.MchApp;
+import com.jeequan.jeepay.core.entity.MchPayPassage;
+import com.jeequan.jeepay.core.entity.PayInterfaceConfig;
+import com.jeequan.jeepay.core.entity.PayOrder;
+import com.jeequan.jeepay.core.exception.BizException;
+import com.jeequan.jeepay.core.utils.StringKit;
+import com.jeequan.jeepay.service.mapper.MchAppMapper;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 /**
+ * <p>
  * 商户应用表 服务实现类
+ * </p>
  *
  * @author [mybatis plus generator]
  * @since 2021-06-15
@@ -42,20 +28,15 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class MchAppService extends ServiceImpl<MchAppMapper, MchApp> {
 
-    @Autowired
-    private PayOrderService payOrderService;
-
-    @Autowired
-    private MchPayPassageService mchPayPassageService;
-
-    @Autowired
-    private PayInterfaceConfigService payInterfaceConfigService;
+    @Autowired private PayOrderService payOrderService;
+    @Autowired private MchPayPassageService mchPayPassageService;
+    @Autowired private PayInterfaceConfigService payInterfaceConfigService;
 
     @Transactional(rollbackFor = Exception.class)
     public void removeByAppId(String appId) {
 
         // 1.查看当前应用是否存在交易数据
-        long payCount = payOrderService.count(PayOrder.gw().eq(PayOrder::getAppId, appId));
+        int payCount = payOrderService.count(PayOrder.gw().eq(PayOrder::getAppId, appId));
         if (payCount > 0) {
             throw new BizException("该应用已存在交易数据，不可删除");
         }
@@ -66,7 +47,8 @@ public class MchAppService extends ServiceImpl<MchAppMapper, MchApp> {
         // 3.删除应用配置的支付参数
         payInterfaceConfigService.remove(PayInterfaceConfig.gw()
                 .eq(PayInterfaceConfig::getInfoId, appId)
-                .eq(PayInterfaceConfig::getInfoType, CS.INFO_TYPE_MCH_APP));
+                .eq(PayInterfaceConfig::getInfoType, CS.INFO_TYPE_MCH_APP)
+        );
 
         // 4.删除当前应用
         if (!removeById(appId)) {
@@ -103,13 +85,13 @@ public class MchAppService extends ServiceImpl<MchAppMapper, MchApp> {
 
         IPage<MchApp> pages = this.page(iPage, wrapper);
 
-        pages.getRecords().stream()
-                .forEach(item -> item.setAppSecret(StringKit.str2Star(item.getAppSecret(), 6, 6, 6)));
+        pages.getRecords().stream().forEach(item -> item.setAppSecret(StringKit.str2Star(item.getAppSecret(), 6, 6, 6)));
 
         return pages;
     }
 
-    public MchApp getOneByMch(String mchNo, String appId) {
+    public MchApp getOneByMch(String mchNo, String appId){
         return getOne(MchApp.gw().eq(MchApp::getMchNo, mchNo).eq(MchApp::getAppId, appId));
     }
+
 }
